Backend/requirements.txt
alembic==1.9.0
anyio==4.6.2.post1
APScheduler==3.11.0
asyncio==3.4.3
asyncpg==0.27.0
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
envparse==0.2.0
exceptiongroup==1.2.2
fastapi==0.88.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.6.2
idna==3.10
iniconfig==2.0.0
load-dotenv==0.1.0
Mako==1.3.6
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pyasn1==0.6.1
pydantic==1.10.19
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.2
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.20.0
virtualenv==20.27.1

falcon~=4.0.2
aiosmtplib~=3.0.2